Refactor httpCarbonMessage
